Move Publishable protocol to own file.
